+++ module/config.js
@@ -531,31 +531,31 @@ ARM5E.covenant.labels = {
     label: "arm5e.sheet.money"
   }
 };
-// influence is the impact the aura has on powers of magic, faery, divine, infernal and mundane respectively
+// influence is the impact the aura has on powers of mundane (ie: none), magic, faery, divine and infernal respectively
 ARM5E.realms = {
   magic: {
     label: "arm5e.sheet.realm.magical",
-    dtype: "String",
-    influence: [1, 0.5, 0, -1, 0]
+    value: 1,
+    influence: [0, 1, 0.5, 0, -1]
   },
   faeric: {
     label: "arm5e.sheet.realm.faeric",
-    dtype: "String",
-    influence: [0.5, 1, 0, -1, 0]
+    value: 2,
+    influence: [0, 0.5, 1, 0, -1]
   },
   divine: {
     label: "arm5e.sheet.realm.divine",
-    dtype: "String",
-    influence: [-3, -4, 1, -5, 0]
+    value: 3,
+    influence: [0, -3, -4, 1, -5]
   },
   infernal: {
     label: "arm5e.sheet.realm.infernal",
-    dtype: "String",
-    influence: [-1, -2, 0, 1, 0]
+    value: 4,
+    influence: [0, -1, -2, 0, 1]
   },
   mundane: {
     label: "arm5e.sheet.realm.mundane",
-    dtype: "String",
+    value: 0,
     influence: [0, 0, 0, 0, 0]
   }
 };